LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-pannernode-interface/distance-inverse-expected.txt

CONSOLE MESSAGE: line 121: TypeError: undefined is not an object (evaluating 'context.startRendering().then')

Harness Error (FAIL), message = TypeError: undefined is not an object (evaluating 'context.startRendering().then')

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [test]  
